docs: suggest "Helped-by" rather than "Thanks-to"
The SubmittingPatches doc suggests adding "tags" in the style of
"Signed-off-by" to commit messages, and suggests some in common usage.
However, based on usage on the Git mailing list (and in particular,
Junio C Hamano's email to René Genz on 23 April), the suggested
"Thanks-to" is not common usage, and should instead be "Helped-by".

Signed-off-by: Adam Dinwoodie <adam@dinwoodie.org>