wt-status.c: Modified status message shown for a parent-less branch
The status message shown for a paren-less branch wasn't so
descriptive and was a little confusing.

The message has been changed to be more descriptive to users in
such a way that it clearly shows the state of the branch / repository.

This allows the users to be reminded that they are about to make
their initial commit which the previous message didn't seem to do
communicate well.

A few alternatives were,

* No commit yet
* Your current branch does not have any commits
* Current branch waiting for initial commit

Signed-off-by: Kaartic Sivaraam <kaarticsivaraam91196@gmail.com>