What's cooking in git.git (Jul 2017, #04; Thu, 13)
Santiago Torres <santiago@nyu.edu> writes:

> On Mon, Jul 17, 2017 at 03:09:44PM -0700, Junio C Hamano wrote:
>> I am not sure if it is merely "if it's even necessary"; if there are
>> two tests running in parallel, with their own separate
>> $TRASH_DIRECTORY, and one of them say "kill the agent" at the
>> beginning, would it affect the other test, depending on the timing?
>
> This shouldn't happen, provided their respective GNUPGHOME env vars
> aren't the same. A gpg-agent process is spawned per GNUPGHOME on the
> versions that I've tested.
>
>>
>> I would imagine that the sockets are kept per GNUPGHOME and they are
>> not going to interfere, so if that is the case, I do not think we
>> mind helping folks with a buggy versions of GnuPG by having a "let's
>> be cautious and kill a leftover agent before starting to test"
>> patch, as long as the reason why we do so is clearly understood and
>> documented.
>
> I double checked the patch/solutions/causes just to be sure I'm not
> doing anything crazy. Here's a v2 of the patch that kills the agent upon
> cleanup rather than startup.

Thanks.

The workaround of killing the lingering agent makes sense to me, but
is test_done the right place to do so?  When "sh t7004-tag.sh -i"
dies in the middle, I do not think the control even reaches the
invocation of that shell function at the end.

The httpd tests seem to do this in start_httpd shell function:

	start_httpd () {
		prepare_httpd
		trap 'code=$?; stop_httpd: (exit $code); die' EXIT
		...

and I can see how this would avoid the "'-i' option can stop the
test script in the middle" issue.  Because the GPG tests do not have
a clear "start agent" step, we'd need to arrange the "trap" in the
case arm in t/lib-gpg.sh where we set GPG prerequisite if we were to
use the same approach.  Or killing a possibly lingering one like your
first patch did may be simpler (we need to make sure &&-chain works
well with "gpgconf --kill all").

Oh, wait, I can run "gpg" just fine, but I do not seem to have
gpgconf.

	$ type gpgconf
	bash: type: gpgconf: not found

The patch may need a bit more cross-version work, it seems.

-- >8 --
From: Santiago Torres <santiago@nyu.edu>
Date: Tue, 18 Jul 2017 13:16:11 -0400
Subject: [RFC/PATCH] t: test-lib: flush gpg agent on cleanup

When running gpg-relevant tests, a gpg-daemon is spawned for each
GNUPGHOME used. This daemon may stay running after the testand cache ile
descriptors for the trash directories, even after the trash directory is
removed. This leads to ENOENT errors when attempting to create files if
tests are run multiple times.

Add a cleanup script to force flushing the gpg-agent when before
removing the trash directory if the test is GPG-relevant.

Helped-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Santiago Torres <santiago@nyu.edu>